+++ tests/word_count/word_count-pthread.c
@@ -173,7 +173,7 @@ void wordcount_splitter(void *data_in, int num_procs)
    /* Barrier, wait for all threads to finish */
    for (i = 0; i < num_procs; i++)
    {
-      int ret_val;
+      long ret_val;
       CHECK_ERROR(pthread_join(tid[i], (void **)(void*)&ret_val) != 0);
 	  CHECK_ERROR(ret_val != 0);
    }
@@ -205,7 +205,7 @@ void wordcount_splitter(void *data_in, int num_procs)
 	   /* Barrier, wait for all threads to finish */
 	   for (i = 0; i < num_threads; i++)
 	   {
-		  int ret_val;
+		  long ret_val;
 		  CHECK_ERROR(pthread_join(tid[i], (void **)(void*)&ret_val) != 0);
 		  CHECK_ERROR(ret_val != 0);
